+++ roulier/carriers/geodis/geodis.py
@@ -1,12 +1,31 @@
 # -*- coding: utf-8 -*-
 """Implementation for Geodis."""
-from geodis_encoder_edi import GeodisEncoderEdi
-from geodis_encoder_ws import GeodisEncoderWs
-from geodis_decoder import GeodisDecoder
-from geodis_transport_ws import GeodisTransportWs
-from geodis_transport_edi import GeodisTransportEdi
+
+from .geodis_encoder_edi import GeodisEncoderEdi
+from .geodis_encoder_ws import GeodisEncoderWs
+from .geodis_encoder_rest_ws import GeodisEncoderRestWs
+from .geodis_decoder_ws import GeodisDecoderWs
+from .geodis_decoder_rest_ws import GeodisDecoderRestWs
+from .geodis_transport_ws import GeodisTransportWs
+from .geodis_transport_edi import GeodisTransportEdi
+from .geodis_transport_rest_ws import GeodisTransportRestWs
+
 from roulier.carrier import Carrier
 from roulier.exception import InvalidAction
+
+# Specifications:
+# rest
+#    usage: tracking
+#    document: GEODIS - GUIDE TECHNIQUE SIC - Nouveau service Zoom - v1.1.docx
+#    date: 2018
+# edi:
+#    usage: deposit slip
+#    document: IFCSUM_D96A_1026_Client_Construit_FR.doc
+#    date: 2012
+# xml:
+#    usage: label and findLocalite
+#    document: GEODIS_Nouvelle eětiquette_GEOLABEL_v FR_V1.9.1.pdf
+#    date: 2016
 
 
 class Geodis(Carrier):
@@ -44,9 +63,15 @@ class Geodis(Carrier):
     def address_validator(self, data, api=False):
         return self._get_ws(data, api, 'findLocalite')
 
+    def get_tracking(self, data, api=False):
+        return self._get_rest_ws(data, api, 'tracking')
+
+    def get_tracking_list(self, data, api=False):
+        return self._get_rest_ws(data, api, 'trackingList')
+
     def _get_ws(self, data, api=False, action=None):
         encoder = GeodisEncoderWs()
-        decoder = GeodisDecoder()
+        decoder = GeodisDecoderWs()
         transport = GeodisTransportWs()
 
         if api:
@@ -60,9 +85,21 @@ class Geodis(Carrier):
             request['infos'],
         )
 
+    def _get_rest_ws(self, data, api=False, action=None):
+        encoder = GeodisEncoderRestWs()
+        decode = GeodisDecoderRestWs()
+        transport = GeodisTransportRestWs()
+        if api:
+            return encoder.api(action=action)
+        request = encoder.encode(data, action)
+        response = transport.send(request)
+        return decode.decode(response, action)
+
     ACTIONS = {
         'label': get_label,
         'findLocalite': address_validator,
         'demandeImpressionEtiquette': get_label,
-        'edi': get_edi
+        'edi': get_edi,
+        'tracking': get_tracking,
+        'trackingList': get_tracking_list,
     }