+++ roulier/carriers/geodis/geodis_api_edi.py
@@ -8,19 +8,19 @@ class GeodisApiEdi(Api):
     def _service(self):
         schema = {
             "depositId": {
-                'type': 'string',
+                'type': 'string', 'coerce': 'accents',
                 'default': '', 'empty': False, 'required': True},
             "depositDate": {
-                'type': 'datetime',
+                'type': 'datetime', 'coerce': 'accents',
                 'default': '', 'empty': False, 'required': True},
             "customerId": {
-                'type': 'string',
+                'type': 'string', 'coerce': 'accents',
                 'default': '', 'empty': False, 'required': True},
             "interchangeSender": {
-                'type': 'string',
+                'type': 'string', 'coerce': 'accents',
                 'default': '', 'empty': False, 'required': True},
             "interchangeRecipient": {
-                'type': 'string',
+                'type': 'string', 'coerce': 'accents',
                 'default': '', 'empty': False, 'required': True},
         }
         return schema
@@ -32,6 +32,9 @@ class GeodisApiEdi(Api):
             'maxlength': 2})
         schema['zip'].update({'required': True, 'empty': False})
         schema['city'].update({'required': True, 'empty': False})
+        for key in schema:
+            if schema[key].get('type', '') == 'string':
+                schema[key].update({'coerce': 'accents'})
         return schema
 
     def _from_address(self):
@@ -41,26 +44,35 @@ class GeodisApiEdi(Api):
         schema['siret'] = {
             'type': 'string', 'required': True,
             'default': '', 'empty': False}
+        for key in schema:
+            if schema[key].get('type', '') == 'string':
+                schema[key].update({'coerce': 'accents'})
         return schema
 
     def _parcel(self):
         weight = GeodisApiWs()._parcel()['weight']
         return {
             'weight': weight,
+            # 'description': 'Barcode of the parcel'
             'barcode': {
                 'type': 'string', 'empty': False, 'default': '',
-                'required': True,
-                'description': 'Barcode of the parcel'}
+                'required': True, 'coerce': 'accents'}
         }
 
+    def _to_address(self):
+        schema = GeodisApiWs()._to_address()
+        for key in schema:
+            if schema[key].get('type', '') == 'string':
+                schema[key].update({'coerce': 'accents'})
+        return schema
+
     def _shipments(self):
-        ws_api = GeodisApiWs()
         v = MyValidator()
         schema = {
             'to_address': {
                 'type': 'dict',
-                'schema': ws_api._to_address(),
-                'default': v.normalized({}, ws_api._to_address())
+                'schema': self._to_address(),
+                'default': v.normalized({}, self._to_address())
             },
             'parcels': {
                 'type': 'list',
@@ -76,18 +88,20 @@ class GeodisApiEdi(Api):
             'productPriority': {
                 'type': 'string', 'default': '', 'empty': True,
                 'required': False,
-                'description': """4219, 1: express, 3: normal speed"""},
+                # 'description': """4219, 1: express, 3: normal speed"""
+                },
             'productOption': {
                 'type': 'string', 'default': '', 'empty': True,
                 'required': False,
-                'description': """7273, like RDV, B2C, BRT, AEX, A2P..."""},
+                # 'description': """7273, like RDV, B2C, BRT, AEX, A2P..."""
+                },
             'notifications': {
                 'type': 'string',
                 'default': GEODIS_ALLOWED_NOTIFICATIONS[0],
                 'allowed': GEODIS_ALLOWED_NOTIFICATIONS,
                 'required': False,
-                'description': """7085 : Notify recipient by
-                    M(ail), S(ms), P(=M+S)"""},
+                # 'description': """7085 : Notify recipient by M(ail), S(ms), P(=M+S)"""
+                },
             'shippingId': {
                 'type': 'string', 'default': '', 'empty': False,
                 'required': True},
@@ -108,7 +122,4 @@ class GeodisApiEdi(Api):
             'agency_address': self._from_address(),
             'from_address': self._from_address(),
         }
-        for schema in schemas:
-            for field in schemas[schema]:
-                schemas[schema][field].update({'coerce': 'accents'})
         return schemas