+++ roulier/carriers/geodis/geodis_decoder_ws.py
@@ -4,7 +4,7 @@ from roulier.codec import Decoder
 from .geodis_common_ws import GEODIS_INFOS
 
 
-class GeodisDecoder(Decoder):
+class GeodisDecoderWs(Decoder):
     """Geodis XML -> Python."""
 
     def decode(self, body, parts, infos):